# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RUTE Travel API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or phone already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out successfully

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    LoginBody:
      type: object
      properties:
        no_whatsapp:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [penumpang, driver]
      required:
        - no_whatsapp
        - password
        - role

    RegisterBody:
      type: object
      properties:
        nama:
          type: string
        no_whatsapp:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [penumpang, driver]
        nik:
          type: string
        kota:
          type: string
        jenis_kendaraan:
          type: string
        plat_nomor:
          type: string
      required:
        - nama
        - no_whatsapp
        - password
        - role

    User:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        no_whatsapp:
          type: string
        role:
          type: string
          enum: [penumpang, driver]
        nik:
          type: string
          nullable: true
        kota:
          type: string
          nullable: true
        jenis_kendaraan:
          type: string
          nullable: true
        plat_nomor:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - nama
        - no_whatsapp
        - role
        - created_at

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user